langchain/llm_interactions/chat_with_gemini.py
from langchain_google_genai import ChatGoogleGenerativeAI
from dotenv import load_dotenv

load_dotenv()

model = ChatGoogleGenerativeAI(model="gemini-2.5-flash")
prompt = "Do you remember me?"

result=model.invoke(prompt)

print(result.content)

# llms are independent to the previous prompt that we write.