langchain/requirements.txt
langchain_community
langchain_google_genai
python-dotenv